feat: add footer and header compontents
