+++ client/src/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import Header from "./component/Header";
+import Footer from "./component/Footer";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 
@@ -8,5 +9,6 @@ root.render(
   <React.StrictMode>
     <Header />
     <h1>Hello React!</h1>
+    <Footer />
   </React.StrictMode>
 );
