--- legacy/layouts/expenses.html
@@ -1,1 +1,0 @@
-<!-- TO BE IMPLEMENTED -->
