--- legacy/layouts/home.html
@@ -1,1 +1,0 @@
-<!-- TO BE IMPLEMENTED -->
